+++ CHANGES.md
@@ -1,6 +1,9 @@
 Unreleased
 =====
 
+1.8.0
+=====
+
 * [#252](https://github.com/serokell/universum/pull/252):
   Remove `Option` re-export. Use `Maybe` instead.